Fixed ping-pong and pin and soon will be channels
