+++ .env
@@ -20,3 +20,12 @@ server_type=login
 
 # Login Channel
 auto_register=true
+
+# World
+world_count=1
+world_0_name=Windia
+world_0_port=8500
+world_0_ribbon=2
+world_0_channels=1
+world_0_event_msg=
+world_0_character_creation_disabled=false
