+++ src/net/handler/login/mod.rs
@@ -1,0 +1,30 @@
+mod login;
+mod pin;
+mod world_select;
+
+use crate::net::client::Client;
+use crate::net::handler::GenericHandler;
+use bytes::Buf;
+use std::sync::{Arc, Mutex};
+
+
+pub struct LoginHandler {}
+
+impl GenericHandler for LoginHandler {
+    fn handle(
+        &self,
+        client: Arc<Mutex<Client>>,
+        buffer: Vec<u8>,
+        _buffer_size: usize,
+    ) -> Option<(Vec<u8>, usize)> {
+        let mut bytes = &buffer.clone()[..];
+
+        match bytes.get_u16_le() {
+            0x1bu16 => login::login(client, &mut bytes),
+            0x03u16 => pin::check_pin_code(client, &mut bytes),
+            0x05u16 => pin::insert_pin_code(client, &mut bytes),
+            0x18u16 => None, // Show Worlds
+            _ => None,
+        }
+    }
+}
